requirements.txt
flask
gunicorn
flask-login
flask-bcrypt
psycopg2-binary